playlists now fetched at app.js level
