+++ src/components/Cards/CardsContainer.js
@@ -9,48 +9,48 @@ import PlaylistModel from '../../models/playlist';
 
 
 const CardsContainer = (props) => { 
-  const playlists = props.playlists
+ 
+  const playlists = props.playlists.playlists
   console.log(playlists)
-  // let list = playlists.map((playlist, index)=>{
-  //   console.log(playlist)
-  // })
-  // let list = playlists.map((playlist, index)=>{
-  //   return <Link to={`/playlist/${playlist._id}`} key={index} >
-  //     <Card 
-  //       { ...playlist }
-  //       />
-  //   </Link> 
-  // })
-    return (
-      <div>
-     
-        {/* <Row>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}>
-            {list[0]}
-          </Col>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}>
-            {list[1]}
-          </Col>
-        </Row>
-        <Row>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}>
-            {list[2]}
-          </Col>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}>
-            {list[3]}
-          </Col>
-        </Row>
-        <Row>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}> 
-            {list[4]}
-          </Col>
-          <Col xs={24} sm={24} md={24} lg={12} xl={12}> 
-            {list[5]}
-          </Col>
-        </Row> */}
-      </div>
-    );
+  let list = []
+  if(playlists){
+    list = playlists.map((playlist, index)=>{
+      return  <Link to={`/playlist/${playlist._id}`} key={index} >
+        <Card 
+          { ...playlist }
+          />
+      </Link> 
+    })
   }
+  return (
+    <div>
+      <Row>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}>
+          {list[0]}
+        </Col>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}>
+          {list[1]}
+        </Col>
+      </Row>
+      <Row>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}>
+          {list[2]}
+        </Col>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}>
+          {list[3]}
+        </Col>
+      </Row>
+      <Row>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}> 
+          {list[4]}
+        </Col>
+        <Col xs={24} sm={24} md={24} lg={12} xl={12}> 
+          {list[5]}
+        </Col>
+      </Row>
+    </div>
+  );
+}
 
 
 export default CardsContainer;
